Pins creation with Transactions.

+++ database/migrations/2020_08_30_132824_create_balances_table.php
@@ -16,10 +16,10 @@
 			Schema::create('balances', function (Blueprint $table) {
 				$table->id();
 				$table->foreignId('user_id');
-				$table->integer('main_balance')->default(0);
-				$table->integer('group_balance')->default(0);
-				$table->integer('mall_balance')->default(0);
-				$table->integer('ad_power_balance')->default(0);
+				$table->decimal('main_balance',20, 10)->default(0);
+				$table->decimal('group_balance',20, 10)->default(0);
+				$table->decimal('mall_balance',20, 10)->default(0);
+				$table->decimal('ad_power_balance',20, 10)->default(0);
 				$table->timestamps();
 
 				$table->foreign('user_id')
